+++ src/app/ApplicationTemplate.Shared/Presentation/Authentication/CreateAccountFormViewModel.cs
@@ -45,12 +45,6 @@ namespace ApplicationTemplate.Presentation
 			set => this.Set(value);
 		}
 
-		public string Password
-		{
-			get => this.Get<string>();
-			set => this.Set(value);
-		}
-
 		public string PhoneNumber
 		{
 			get => this.Get<string>();
@@ -79,66 +73,6 @@ namespace ApplicationTemplate.Presentation
 		{
 			get => this.Get<bool>();
 			set => this.Set(value);
-		}
-
-		public bool? PasswordHasMinimumLength
-		{
-			get => this.GetFromObservable(ObservePasswordHasMinimumLength(), initialValue: null);
-		}
-
-		public bool? PasswordHasNumber
-		{
-			get => this.GetFromObservable(ObservePasswordHasNumber(), initialValue: null);
-		}
-
-		public bool? PasswordHasUppercase
-		{
-			get => this.GetFromObservable(ObservePasswordHasUppercase(), initialValue: null);
-		}
-
-		private IObservable<bool?> ObservePasswordHasMinimumLength()
-		{
-			return this.GetProperty(x => x.Password)
-				.Observe()
-				.Select<string, bool?>(password =>
-				{
-					if (password.IsNullOrEmpty())
-					{
-						return null;
-					}
-
-					return password.Length >= PresentationConstants.PasswordMinLength;
-				});
-		}
-
-		private IObservable<bool?> ObservePasswordHasNumber()
-		{
-			return this.GetProperty(x => x.Password)
-				.Observe()
-				.Select<string, bool?>(password =>
-				{
-					if (password.IsNullOrEmpty())
-					{
-						return null;
-					}
-
-					return password.Any(char.IsDigit);
-				});
-		}
-
-		private IObservable<bool?> ObservePasswordHasUppercase()
-		{
-			return this.GetProperty(x => x.Password)
-				.Observe()
-				.Select<string, bool?>(password =>
-				{
-					if (password.IsNullOrEmpty())
-					{
-						return null;
-					}
-
-					return password.Any(char.IsUpper);
-				});
 		}
 	}
