+++ src/app/ApplicationTemplate.UWP/Views/Content/Authentication/CreateAccountPage.xaml
@@ -157,14 +157,14 @@
 					</local:DataValidationView>
 
 				<!-- Password -->
-				<local:DataValidationView Model="{Binding Form}"
+					<local:DataValidationView Model="{Binding PasswordForm}"
 										  PropertyName="Password"
 										  Margin="42,0"
 										  Background="{StaticResource MaterialSurfaceBrush}">
 					<PasswordBox x:Name="PasswordField"
 								 x:Uid="CreateAccount_Password"
 								 PlaceholderText="Choose a password"
-								 Password="{Binding Form.Password, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
+								 Password="{Binding PasswordForm.Password, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
 								 ios:ReturnKeyType="Done"
 								 android:ImeOptions="Done"
 								 ue:PasswordBoxBehavior.DismissKeyboardOnEnter="True" />
@@ -183,19 +183,19 @@
 							<ColumnDefinition Width="*" />
 						</Grid.ColumnDefinitions>
 
-						<u:PathControl x:Name="XHasMinLengthPathControl"
+						<u:PathControl x:Name="HasMinLengthPathControl"
 										   VerticalAlignment="Center"
-										   Style="{Binding Form.PasswordHasMinimumLength, Converter={StaticResource NullableBoolToIconStyle}}"
+										   Style="{Binding PasswordForm.PasswordHasMinimumLength, Converter={StaticResource NullableBoolToIconStyle}}"
 										   Margin="9,4,9,4" />
 
-						<u:PathControl x:Name="XHasNumberPathControl"
-										   Style="{Binding Form.PasswordHasNumber, Converter={StaticResource NullableBoolToIconStyle}}"
+						<u:PathControl x:Name="HasNumberPathControl"
+										   Style="{Binding PasswordForm.PasswordHasNumber, Converter={StaticResource NullableBoolToIconStyle}}"
 										   VerticalAlignment="Center"
 										   Grid.Row="1"
 										   Margin="9,4,9,4" />
 
-						<u:PathControl x:Name="XHasUppercasePathControl"
-										   Style="{Binding Form.PasswordHasUppercase, Converter={StaticResource NullableBoolToIconStyle}}"
+						<u:PathControl x:Name="HasUppercasePathControl"
+										   Style="{Binding PasswordForm.PasswordHasUppercase, Converter={StaticResource NullableBoolToIconStyle}}"
 										   VerticalAlignment="Center"
 										   Grid.Row="2"
 										   Margin="9,4,9,4" />
