+++ src/app/ApplicationTemplate.UWP/Views/Content/Authentication/ResetPasswordPage.xaml
@@ -21,7 +21,6 @@
 		</ResourceDictionary>
 	</Page.Resources>
 
-
 	<Grid Background="{StaticResource MaterialBackgroundBrush}">
 
 		<!-- Background -->
@@ -63,14 +62,14 @@
 							   HorizontalAlignment="Center" />
 
 					<!-- Password -->
-					<local:DataValidationView Model="{Binding Password}"
+					<local:DataValidationView Model="{Binding PasswordForm}"
 											  PropertyName="Password"
 											  Margin="42,12,42,4"
 											  Background="{StaticResource MaterialSurfaceBrush}">
 						<PasswordBox x:Name="PasswordField"
 									 x:Uid="ResetPassword_Password"
 									 PlaceholderText="Choose a new password"
-									 Password="{Binding Password, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
+									 Password="{Binding PasswordForm.Password, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
 									 ios:ReturnKeyType="Next"
 									 android:ImeOptions="Next"
 									 ue:PasswordBoxBehavior.DismissKeyboardOnEnter="True" />
@@ -89,19 +88,19 @@
 							<ColumnDefinition Width="*" />
 						</Grid.ColumnDefinitions>
 
-						<u:PathControl x:Name="XHasMinLengthPathControl"
+						<u:PathControl x:Name="HasMinLengthPathControl"
 									   VerticalAlignment="Center"
-									   Style="{Binding PasswordHasMinimumLength, Converter={StaticResource NullableBoolToIconStyle}}"
+									   Style="{Binding PasswordForm.PasswordHasMinimumLength, Converter={StaticResource NullableBoolToIconStyle}}"
 									   Margin="9,4,9,4" />
 
-						<u:PathControl x:Name="XHasNumberPathControl"
-									   Style="{Binding PasswordHasNumber, Converter={StaticResource NullableBoolToIconStyle}}"
+						<u:PathControl x:Name="HasNumberPathControl"
+									   Style="{Binding PasswordForm.PasswordHasNumber, Converter={StaticResource NullableBoolToIconStyle}}"
 									   VerticalAlignment="Center"
 									   Grid.Row="1"
 									   Margin="9,4,9,4" />
 
-						<u:PathControl x:Name="XHasUppercasePathControl"
-									   Style="{Binding PasswordHasUppercase, Converter={StaticResource NullableBoolToIconStyle}}"
+						<u:PathControl x:Name="HasUppercasePathControl"
+									   Style="{Binding PasswordForm.PasswordHasUppercase, Converter={StaticResource NullableBoolToIconStyle}}"
 									   VerticalAlignment="Center"
 									   Grid.Row="2"
 									   Margin="9,4,9,4" />
